revert filter incompatibility check.
The previous implementation viewed the namespace and prefix query
parameters as a type of filter, and it also attempted to short-circuit
the filter operation by usig these params instead of the go-bexpr
filter.

This turned out to be not ideal because the value for namespace can be
set in other places instead of the direct request (like globally at the
API client), and using a filter equality search is not the same as using
prefix, so neither can be considered a filter.

Short-circuiting can also lead to unexpected results because is not
possible to tell if the user specified a namespace at the RPC layer,
since it will be set to default by the HTTP server if empty.

If the user explicitly set `namespace=default` and use the filter
`Namespace == "other"`, the requested namespace would be overriden.